requirements.txt
PyPDF2
mysql-connector-python
python-dotenv